+++ application/controllers/Mtce.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: daniel
+ * Date: 2017-10-19
+ * Time: 3:18 PM
+ */
+
+class Mtce extends Application {
+
+    public function index() {
+        $this->data['pagetitle'] = 'TODO List Maintenance';
+        $tasks = $this->tasks->all(); // get all the tasks
+
+        // substitute the status name
+        foreach ($tasks as $task)
+            if (!empty($task->status))
+                $task->status = $this->app->status($task->status);
+
+        // build the task presentation output
+        $result = '';   // start with an empty array
+        foreach ($tasks as $task)
+            $result .= $this->parser->parse('oneitem',(array)$task,true);
+
+        // and then pass them on
+        $this->data['display_tasks'] = $result;
+        $this->data['pagebody'] = 'itemlist';
+        $this->render();
+    }
+}